Move common functions to utilities, fix dom plugin
