+++ sbt-scalajs-esbuild-dom/src/sbt-test/sbt-scalajs-esbuild-dom/basic-dom-project/build.sbt
@@ -1,4 +1,4 @@
-enablePlugins(ScalaJSEsbuildPlugin)
+enablePlugins(ScalaJSEsbuildDomPlugin)
 
 scalaVersion := "2.13.8"
