+++ sbt-scalajs-esbuild-dom/src/sbt-test/sbt-scalajs-esbuild-dom/basic-dom-project/esbuild/package.json
