+++ sbt-scalajs-esbuild-dom/src/sbt-test/sbt-scalajs-esbuild-dom/basic-dom-project/project/plugins.sbt
@@ -20,7 +20,9 @@ if (sourcePlugins.nonEmpty) {
     sys.error("'plugin.version' environment variable is not set")
   )
   Seq(
-    addSbtPlugin("me.ptrdom" % "sbt-scalajs-esbuild" % scalaJSEsbuildVersion)
+    addSbtPlugin(
+      "me.ptrdom" % "sbt-scalajs-esbuild-dom" % scalaJSEsbuildVersion
+    )
   )
 }
