+++ sbt-scalajs-esbuild/src/main/scala/scalajsesbuild/ScalaJSEsbuildPlugin.scala
@@ -59,152 +59,6 @@ object ScalaJSEsbuildPlugin extends AutoPlugin {
 
   import autoImport._
 
-  // TODO consider using sbt FileChanges instead
-  sealed trait ChangeStatus
-  object ChangeStatus {
-    case object Pristine extends ChangeStatus
-    case object Dirty extends ChangeStatus
-
-    implicit class ChangeStatusOps(changeStatus: ChangeStatus) {
-      def combine(other: ChangeStatus): ChangeStatus = {
-        (changeStatus, other) match {
-          case (Pristine, Pristine) =>
-            Pristine
-          case _ => Dirty
-        }
-      }
-    }
-  }
-
-  private[scalajsesbuild] sealed trait Stage
-  private[scalajsesbuild] object Stage {
-    case object FullOpt extends Stage
-    case object FastOpt extends Stage
-  }
-
-  private[scalajsesbuild] def jsFileNames(report: Report) = {
-    report match {
-      case report: unstable.ReportImpl =>
-        val jsFileNames = report.publicModules
-          .map { publicModule =>
-            publicModule.jsFileName
-          }
-        jsFileNames
-      case unhandled =>
-        sys.error(s"Unhandled report type [$unhandled]")
-    }
-  }
-
-  private def esbuildLoaders = {
-    // taken from Vite's KNOWN_ASSET_TYPES constant
-    Seq(
-      // images
-      "png",
-      "jpe?g",
-      "jfif",
-      "pjpeg",
-      "pjp",
-      "gif",
-      "svg",
-      "ico",
-      "webp",
-      "avif",
-
-      // media
-      "mp4",
-      "webm",
-      "ogg",
-      "mp3",
-      "wav",
-      "flac",
-      "aac",
-
-      // fonts
-      "woff2?",
-      "eot",
-      "ttf",
-      "otf",
-
-      // other
-      "webmanifest",
-      "pdf",
-      "txt"
-    ).map(assetType => s"'.$assetType': 'file'")
-      .mkString(",")
-  }
-
-  private[scalajsesbuild] def esbuildOptions(
-      entryPoints: Seq[String],
-      outdir: String,
-      hashOutputFiles: Boolean,
-      minify: Boolean
-  ) = {
-    val entryPointsFn: Seq[String] => String = _.map(escapePathString)
-      .mkString(",")
-    val outdirFn: String => String = escapePathString
-
-    Seq(
-      s"""
-         |  entryPoints: [${entryPointsFn(entryPoints)}],
-         |  bundle: true,
-         |  outdir: '${outdirFn(outdir)}',
-         |  loader: { $esbuildLoaders },
-         |  metafile: true,
-         |  logOverride: {
-         |    'equals-negative-zero': 'silent',
-         |  },
-         |  logLevel: "info",
-         |""".stripMargin.some,
-      if (hashOutputFiles) {
-        """
-          |  entryNames: '[name]-[hash]',
-          |  assetNames: '[name]-[hash]',
-          |""".stripMargin.some
-      } else None,
-      if (minify) {
-        """
-          |  minify: true,
-          |""".stripMargin.some
-      } else None
-    ).flatten.mkString
-  }
-
-  private[scalajsesbuild] def escapePathString(pathString: String) =
-    pathString.replace("\\", "\\\\")
-
-  private def processFileChanges(
-      fileChanges: FileChanges,
-      sourceDirectory: File,
-      targetDirectory: File
-  ): Unit = {
-    def toTargetFile(
-        sourcePath: Path,
-        sourceDirectory: File,
-        targetDirectory: File
-    ): File = {
-      new File(
-        sourcePath.toFile.getAbsolutePath.replace(
-          sourceDirectory.getAbsolutePath,
-          targetDirectory.getAbsolutePath
-        )
-      )
-    }
-
-    (fileChanges.created ++ fileChanges.modified)
-      .foreach { path =>
-        IO.copyFile(
-          path.toFile,
-          toTargetFile(path, sourceDirectory, targetDirectory)
-        )
-      }
-
-    fileChanges.deleted.foreach { path =>
-      IO.delete(
-        toTargetFile(path, sourceDirectory, targetDirectory)
-      )
-    }
-  }
-
   override lazy val projectSettings: Seq[Setting[_]] = Seq(
     scalaJSLinkerConfig ~= {
       _.withModuleKind(ModuleKind.ESModule)
@@ -336,10 +190,8 @@ object ScalaJSEsbuildPlugin extends AutoPlugin {
     perScalaJSStageSettings(Stage.FullOpt)
 
   private def perScalaJSStageSettings(stage: Stage): Seq[Setting[_]] = {
-    val (stageTask, stageTaskWrapper) = stage match {
-      case Stage.FastOpt => (fastLinkJS, esbuildFastLinkJSWrapper)
-      case Stage.FullOpt => (fullLinkJS, esbuildFullLinkJSWrapper)
-    }
+    val (stageTask, stageTaskWrapper) =
+      (stage.stageTask, stage.stageTaskWrapper)
 
     Seq(
       stageTask / esbuildCompile := {
